Update NodeMessage.ts
I've been using Veza for a while for my Discord bot because it's used by Kurasuta which is my sharding manager. Ever since I started using Veza I noticed the issue that it sometimes tries to write on the Socket when it cannot. I've reported this on 12/07/2020 in Veza's support server, I showed my temporary fix. There have been no changes since then and it's starting to annoy me that I have to change it manually every time otherwise it crashes my entire process when I restart separate clusters. This change should only be temporary until a real fix has been implemented but it will save me and other bot developers a lot of time and effort to constantly update this line.  Although it falls silent when it should at least throw an error I've been using it like this in production with 98k servers without any issues for the past months.